ipv6支持

+++ libuv_cpp11/uv/TcpConnection.cpp
@@ -108,7 +108,7 @@ int TcpConnection::write(const char* buf,unsigned int size,AfterWriteCallback ca
     return rst;
 }
 
-void TcpConnection::writeInLoop(const char* buf,unsigned int size,AfterWriteCallback callback)
+void TcpConnection::writeInLoop(const char* buf,ssize_t size,AfterWriteCallback callback)
 {
     Async<struct write_arg_t>* async = new Async<struct write_arg_t>(loop, 
     std::bind([this](Async<struct write_arg_t>* handle, struct write_arg_t * data)
